requirements.txt
dj-database-url==0.4.2
Django==2.0.1
django-heroku==0.2.0
django-markdown==0.8.4
gunicorn==19.7.1
Markdown==2.6.11
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
whitenoise==3.3.1
